rapid/pyproject.toml
[project]
name = "rapid-clone-bot"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.13.1",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.2",
    "watchfiles>=0.22.0",
]

[tool.uv]
package = false
